Making like the Tytalus and driving the Massassa out
