+++ data-sample/stats/flaws.ts
@@ -1,4 +1,4 @@
-import { IMStat } from "../../../@types/index.ts";
+import { IMStat } from "../../src/@types/index.ts";
 
 export const flaws: IMStat[] = [
   {
