+++ src/commands/sheet.ts
@@ -154,241 +154,6 @@ const skills = async (obj: Obj) => {
   return output;
 };
 
-const advantages = (obj: Obj) => {
-  let output = `${divider("Backgrounds", "%cr-%cn", 26)}${
-    divider(
-      "Advantages",
-      "%cr-%cn",
-      26,
-    )
-  }${divider("Flaws", "%cr-%cn", 26)}\n`;
-
-  const backgrounds = obj.data?.stats
-    ?.filter((s) => s.type === "background")
-    .map(async (s) => formatStat(s.name, await getStat(obj.dbobj, s.name))) ||
-    [];
-
-  const merits = obj.data?.stats
-    ?.filter((s) => s.type === "merit")
-    .map(async (s) => formatStat(s.name, await getStat(obj.dbobj, s.name))) ||
-    [];
-  const flaws = obj.data?.stats
-    ?.filter((s) => s.type === "flaw")
-    .map(async (s) => formatStat(s.name, await getStat(obj.dbobj, s.name))) ||
-    [];
-
-  const max = Math.max(backgrounds.length, merits.length, flaws.length);
-  const totalBackgrounds = [];
-  const totalMerits = [];
-  const totalFlaws = [];
-
-  totalBackgrounds.push(
-    ...backgrounds,
-    ...Array(max - backgrounds.length).fill(" ".repeat(24)),
-  );
-  totalMerits.push(
-    ...merits,
-    ...Array(max - merits.length).fill(" ".repeat(24)),
-  );
-  totalFlaws.push(...flaws, ...Array(max - flaws.length).fill(" ".repeat(24)));
-
-  for (let i = 0; i < max; i++) {
-    output += ` ${totalBackgrounds[i]}  ${totalMerits[i]}  ${totalFlaws[i]}\n`;
-  }
-  if (max) {
-    return output;
-  }
-
-  return "";
-};
-
-const disciplines = async (obj: Obj) => {
-  const splat = await getStat(obj.dbobj, "splat");
-
-  const totalDisciplines: any[] = [];
-
-  const disciplines = obj.data?.stats
-    ?.filter((s) => s.type === "discipline")
-    .sort((a, b) => a.name.localeCompare(b.name)) || [];
-
-  // split the disciplines into two columns. using a columns array.
-
-  const columns: any[][] = [[], []];
-  for (let i = 0; i < disciplines.length; i++) {
-    columns[i % 2].push(disciplines[i]);
-  }
-
-  columns[0] = columns[0].sort((a, b) => a.name.localeCompare(b.name));
-  columns[1] = columns[1].sort((a, b) => a.name.localeCompare(b.name));
-
-  for (const col of columns) {
-    const colDisciplines = [];
-    for (const stat of col) {
-      colDisciplines.push(
-        formatStat(stat.name, await getStat(obj.dbobj, stat.name), 37),
-      );
-
-      for (
-        const s of obj.data?.stats
-          ?.filter((s) => s.type === stat.name)
-          .sort((a, b) => a.value - b.value) || []
-      ) {
-        colDisciplines.push(
-          "   " + formatStat(s.name, await getStat(obj.dbobj, s.name), 34),
-        );
-      }
-
-      colDisciplines.push(" ".repeat(37));
-    }
-    totalDisciplines.push(colDisciplines);
-  }
-
-  let output = divider("Disciplines") + "%r ";
-  const max = Math.max(totalDisciplines[0].length, totalDisciplines[1].length);
-  totalDisciplines[0].push(
-    ...Array(max - totalDisciplines[0].length).fill(" ".repeat(37)),
-  );
-  totalDisciplines[1].push(
-    ...Array(max - totalDisciplines[1].length).fill(" ".repeat(37)),
-  );
-
-  output += totalDisciplines[0]
-    .map((d: any, i: number) => ` ${d}  ${totalDisciplines[1][i]}`)
-    .join("\n")
-    .trim();
-
-  if (!max) return "";
-  return output + "%r";
-};
-
-const other = async (obj: Obj) => {
-  const splat = await getStat(obj.dbobj, "splat");
-  let output = "";
-
-  const other = allStats.filter(
-    (stat) =>
-      stat.type === "other" && (stat.splat?.includes(splat) || !stat.splat),
-  );
-
-  let totalOther = [];
-
-  for (const stat of other) {
-    totalOther.push(formatStat(stat.name, await getStat(obj.dbobj, stat.name)));
-  }
-  totalOther = totalOther.sort((a, b) => a.localeCompare(b));
-  output += "%cr-%cn".repeat(78);
-
-  for (let i = 0; i < totalOther.length; i++) {
-    if (i % 3 === 0) {
-      output += "%r%b";
-    }
-    output += totalOther[i] + "  ";
-  }
-
-  return output + "%r";
-};
-
-const calculateDamage = (
-  superficial: number,
-  aggravated: number,
-  maxBoxes: number,
-  characterType: string, // Assuming characterType can be "mortal", "ghoul", or others
-) => {
-  let damageBoxes = Array(maxBoxes).fill("[ ]");
-  let status = "";
-
-  // Apply Aggravated damage
-  for (let i = 0; i < aggravated && i < maxBoxes; i++) {
-    damageBoxes[i] = "[X]";
-  }
-
-  // Apply Superficial damage in remaining boxes if available
-  for (let i = 0; i < maxBoxes; i++) {
-    if (damageBoxes[i] === "[ ]" && superficial > 0) {
-      damageBoxes[i] = "[/]";
-      superficial--;
-    }
-  }
-
-  // Upgrade Superficial to Aggravated if needed
-  if (superficial > 0) {
-    for (let i = 0; i < maxBoxes && superficial > 0; i++) {
-      if (damageBoxes[i] === "[/]") {
-        damageBoxes[i] = "[X]";
-        superficial--;
-      }
-    }
-  }
-
-  // Check for Impaired or Incapacitated status
-  const filledBoxes = damageBoxes.reduce(
-    (acc, val) => acc + (val !== "[ ]" ? 1 : 0),
-    0,
-  );
-
-  if (filledBoxes >= maxBoxes) {
-    // Special rule for mortal and ghoul characters when they reach impaired
-    if (
-      (characterType === "mortal" || characterType === "ghoul") &&
-      aggravated < maxBoxes
-    ) {
-      status = "Incapacitated";
-    } else if (aggravated === maxBoxes || superficial > 0) {
-      // All boxes are aggravated or track is overflowing
-      status = "Incapacitated";
-    } else {
-      status = "Impaired";
-    }
-  }
-
-  return { damageBoxes, status };
-};
-
-const displayDamageTrack = async (obj: IDBOBJ, type: string) => {
-  let output = "";
-  let splat = await getStat(obj, "splat");
-  const superficial = parseInt(obj.data?.damage[type].superficial || 0);
-  const aggravated = parseInt(obj.data?.damage[type].aggravated || 0);
-  let maxBoxes;
-
-  if (type === "physical") {
-    maxBoxes = (await getStat(obj, "stamina")) + 3;
-  } else if (type === "mental") {
-    const composure = await getStat(obj, "composure");
-    const resolve = await getStat(obj, "resolve");
-    maxBoxes = composure + resolve;
-  } else {
-    throw new Error("Invalid damage type");
-  }
-
-  const { damageBoxes, status } = calculateDamage(
-    superficial,
-    aggravated,
-    maxBoxes,
-    splat,
-  );
-  let trackLabel = type === "physical" ? "Health:    " : "Willpower: ";
-
-  if (status) {
-    output += ` ${trackLabel}%ch%cr${damageBoxes.join("")}%cn`;
-    output += ` %ch%cr(${status})%cn`;
-  } else {
-    output += ` ${trackLabel}${damageBoxes.join("")} `;
-  }
-
-  return output;
-};
-
-const health = async (obj: IDBOBJ) => {
-  let output = divider("Character Status") + "\n";
-
-  output += await displayDamageTrack(obj, "physical");
-  output += "%r";
-  output += await displayDamageTrack(obj, "mental");
-  output += "%r"; // Include additional tracks or information as needed
-
-  return output;
-};
 /*sheet
  ===========================[  Sheet for: Kumakun  ]===========================
  Full Name:                             Concept:
@@ -409,12 +174,6 @@ const health = async (obj: IDBOBJ) => {
  Melee..................0  Persuasion.............0  Politics...............0
  Stealth................0  Streetwise.............0  Science................0
  Survival...............0  Subterfuge.............0  Technology.............0
------------------------------------ Health -----------------------------------
- Health: [ ][ ][ ][ ][ ][ ][ ][ ]
-------------------------------------------------------------------------------
- Health.................0  Willpower..............0  Humanity...............0
- Blood Potency..........1  Blood Pool.............0  Experience.............0
-==============================================================================
  */
 export default () => {
   addCmd({
@@ -441,24 +200,7 @@ export default () => {
       output += await bio(tarObj);
       output += await attributes(tarObj);
       output += await skills(tarObj);
-      output += await advantages(tarObj);
-      output += await disciplines(tarObj);
-      output += await health(tarObj);
-      output += await other(tarObj);
       output += footer();
-
-      if (await getStat(tarObj.dbobj, "splat")) {
-        send([ctx.socket.id], output);
-      } else {
-        if (tarObj.dbref === en.dbref) {
-          send(
-            [ctx.socket.id],
-            "%chGame>%cn You have no splat set. See: %ch+help splat%cn",
-          );
-        } else {
-          send([ctx.socket.id], "%chGame>%cn That character has no splat set.");
-        }
-      }
     },
   });
 };
