+++ src/services/characters/index.ts
@@ -1,11 +1,6 @@
 import { attributes } from "./stats/attributes.ts";
-import { backgrounds } from "./stats/backgrounds.ts";
 import { bio } from "./stats/bio.ts";
-import { flaws } from "./stats/flaws.ts";
-import { merits } from "./stats/merits.ts";
 import { skills } from "./stats/skills.ts";
-import { disciplines } from "./stats/disciplines.ts";
-import { other } from "./stats/other.ts";
 
 export * from "./attributes.ts";
 export * from "./setStat.ts";
@@ -13,13 +8,12 @@ export * from "./getStats.ts";
 export * from "./formatValue.ts";
 export * from "./statObj.ts";
 
-export const allStats = [
+export let allStats = [
   ...bio,
   ...attributes,
   ...skills,
-  ...merits,
-  ...flaws,
-  ...backgrounds,
-  ...disciplines,
-  ...other,
 ];
+
+export function setAllStats(newStats: array) : array {
+  allStats = newStats;
+}
