+++ src/services/characters/stats/bio.ts
@@ -1,0 +1,19 @@
+import { IMStat } from "../../../@types/index.ts";
+
+export const bio: IMStat[] = [
+  {
+    name: "full name",
+    values: [],
+    type: "bio",
+  },
+  {
+    name: "concept",
+    values: [],
+    type: "bio",
+  },
+  {
+    name: "Birth Date",
+    values: [],
+    type: "bio",
+  },
+];
